refactor(components): move files to folders - web components to own folder - use shared folder - tailwind to styles folder - mutation, temporal and other utils to utils folder
