+++ components/index.html
@@ -6,7 +6,7 @@
         <title>Components Demo</title>
         <script src="./node_modules/@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
         <script src="./node_modules/lit/polyfill-support.js"></script>
-        <script type="module" src="./src/components/app.ts"></script>
+        <script type="module" src="src/preact-components/app.ts"></script>
         <script type="module" src="src/components/prevalenceOverTime/prevalence-over-time.ts"></script>
         <script type="module" src="src/preact-components/input/location-filter-component.tsx"></script>
         <script type="module" src="src/components/dateRangeSelector/date-range-selector.ts"></script>
