+++ components/src/operator/FetchAggregatedOperator.ts
@@ -1,7 +1,7 @@
 import { Operator } from './Operator';
 import { Dataset } from './Dataset';
 import { LapisFilter } from '../types';
-import { mapLapisFilterToUrlParams } from '../utils';
+import { mapLapisFilterToUrlParams } from '../utils/utils';
 
 export class FetchAggregatedOperator<Fields> implements Operator<Fields & { count: number }> {
     constructor(
