+++ components/src/preact/shared/charts/getYAxisScale.ts
