+++ components/src/preact/shared/table/formatProportion.ts
@@ -1,0 +1,3 @@
+export const formatProportion = (proportion: number) => {
+    return `${(proportion * 100).toFixed(2)}%`;
+};
