+++ components/src/styles/tailwind.d.ts
