+++ components/src/web-components/display/mutation-comparison-component.stories.ts
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/web-components';
 import { html } from 'lit';
 import './mutation-comparison-component';
-import '../../components/app';
+import '../app';
 import { LAPIS_URL, NUCLEOTIDE_INSERTIONS_ENDPOINT, NUCLEOTIDE_MUTATIONS_ENDPOINT } from '../../constants';
 import { MutationComparisonProps } from '../../preact/mutationComparison/mutation-comparison';
 import nucleotideMutationsSomeVariant from '../../preact/mutationComparison/__mockData__/nucleotideMutationsSomeVariant.json';
